pyproject.toml
[project]
name = "django-entraid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.6",
    "djangorestframework>=3.15.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.6",
]
